chore: rename python package hive → starter
